requirements.txt
# Core Django - Use LTS version for stability
Django==4.2.7
djangorestframework==3.14.0

# Celery for task queue
celery==5.3.4
redis==5.0.1

# Data processing
pandas==2.1.4
numpy==1.26.0
requests==2.31.0

# Visualization - Use compatible versions
plotly==5.17.0
dash==2.14.2

# Database
psycopg2-binary==2.9.7

# Environment and deployment
python-dotenv==1.0.0
django-cors-headers==4.3.1
gunicorn==21.2.0
flower==2.0.1
django-environ==0.10.0
dj-database-url==2.1.0
whitenoise==6.6.0

# Build tools
setuptools==68.2.2
wheel==0.41.2
